backend/requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-text-splitters
pypdf
fastapi
uvicorn
python-dotenv
pinecone
langchain-pinecone
langchain-tavily
langchain-openai
requests
plotly